Bug fix and enhancement
Allow adding a specific color's track to any neighboring adjacent cell to where that color resides regardless of which color placed the track that the game token is on. Also increased the number of tracks to: 4 cross tracks, 12 bend/elbow tracks, and 14 straight tracks. In testing discovered that players run out of tracks and more are neede.